Add ground truth preparation

+++ libs/rag_eval/src/rag_eval/data.py
@@ -12,6 +12,7 @@ data_dir = fs.info("data", detail=False)["name"]
 ground_truth_dir = posixpath.join(data_dir, "ground_truth")
 answers_dir = posixpath.join(data_dir, "answers")
 evaluation_results_dir = posixpath.join(data_dir, "evaluation_results")
+questions_dir = posixpath.join(data_dir, "questions")
 
 
 @dataclass(frozen=True)
@@ -51,6 +52,8 @@ def get_answers_datasets_list() -> list[DatasetPath]:
 def get_evaluation_results_datasets_list() -> list[DatasetPath]:
     return list_data_files(evaluation_results_dir)
 
+def get_questions_datasets_list() -> list[DatasetPath]:
+    return list_data_files(questions_dir)
 
 @st.cache_data
 def read_dataset(dataset_path: DatasetPath):
@@ -66,13 +69,34 @@ def read_ground_truth(dataset_path: DatasetPath):
 def read_answers(dataset_path: DatasetPath):
     return read_dataset(dataset_path)
 
+@st.cache_data
+def read_questions(dataset_path: DatasetPath):
+    return read_dataset(dataset_path)
 
-def write_evaluation_results(name: str, data: pd.DataFrame) -> DatasetPath:
-    fs.makedirs(evaluation_results_dir, exist_ok=True)
+
+def write_dataset(parent_dir: str, name: str, data: pd.DataFrame, rewrite: bool = False) -> DatasetPath:
+    fs.makedirs(parent_dir, exist_ok=True)
     if not name.endswith(".parquet"):
         name = f"{name}.parquet"
     dataset_path = DatasetPath.create_as_relative(
-        posixpath.join(evaluation_results_dir, name), evaluation_results_dir
+        posixpath.join(parent_dir, name), parent_dir
     )
+    if not rewrite and fs.exists(dataset_path.path):
+        raise FileExistsError(f"File {dataset_path} already exists")
+
     data.to_parquet(dataset_path.path)
     return dataset_path
+
+
+def write_evaluation_results(name: str, data: pd.DataFrame) -> DatasetPath:
+    return write_dataset(evaluation_results_dir, name, data, rewrite=True)
+
+
+def write_questions(name: str, data: pd.DataFrame, rewrite: bool=False) -> DatasetPath:
+    return write_dataset(questions_dir, name, data, rewrite)
+
+def write_answers(name: str, data: pd.DataFrame, rewrite: bool=False) -> DatasetPath:
+    return write_dataset(answers_dir, name, data, rewrite)
+
+def write_ground_truth(name: str, data: pd.DataFrame, rewrite: bool=False) -> DatasetPath:
+    return write_dataset(ground_truth_dir, name, data, rewrite)
